Renamed component DropdownSelectbox to DropdownCheckbox, refactored code
